app/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
streamlit==1.22.0
pillow==9.5.0
numpy==1.24.3
pandas==2.0.2
requests==2.31.0
python-multipart==0.0.6
torch==2.0.1
torchvision==0.15.2 